Import py26-libtorrent-rasterbar-0.15.0 as wip/py-libtorrent-rasterbar.
libtorrent is a feature complete C++ bittorrent implementation
focusing on efficiency and scalability. It runs on embedded devices as
well as desktops. It boasts a well documented library interface that
is easy to use. It comes with a simple bittorrent client demonstrating
the use of the library.
This package provides python binding library.