Initial import of gnome-user-share-2.26.0.
gnome-user-share is a small package that binds together various
free software projects to bring easy to use user-level file sharing
to the masses.

The program is meant to run in the background when the user is
logged in, and when file sharing is enabled a webdav server is
started that shares the $HOME/Public folder. The share is then
published to all computers on the local network using mDNS/rendezvous,
so that it shows up in the Network location in Gnome.

The dav server used is apache, so you need that installed. Avahi
or Howl is used for mDNS support, so you need to have that installed
and mDNSResolver running.

TODO:

Package is fine, but installation notes say:

Do note Apache 2.2 and mod_dnssd 0.6 are required to use gnome-user-share.

Currently unclear how to set this up from pkgsrc.