Import py26-polybori-0.6-3 as wip/py-ploybori.
The core of PolyBoRi is a C++ library, which provides high-level data types
for Boolean polynomials and monomials, exponent vectors, as well as for the
underlying polynomial rings and subsets of the powerset of the Boolean
variables. As a unique approach, binary decision diagrams are used as
internal storage type for polynomial structures. On top of this C++-library
we provide a Python interface. This allows parsing of complex polynomial
systems, as well as sophisticated and extendable strategies for Groebner
base computation. PolyBoRi features a powerful reference implementation for
Groebner basis computation.