Import py26-ssh2-0.6.9 as wip/py-ssh2.
ssh4py is a Python wrapper around libssh2. This package removes methods
which require the private API files from libssh2 (channel.getSocket,
ssh.getCompressionMethods) and a method which was not exported, and
which requires a function that doesn't exist in libssh2 (channel.pollWrite).

This package requires a bit more polishing I think. There are still hard-coded
paths that should not be required (to be tested on Darwin). All patches should
be commented.