Import thwab-2.2.3 as wip/thwab.
Thwab is an Electronic Encyclopedia System that is international, portable,
simple, extensible (as one could add more books and articles), fast, caching,
smart and maybe at some level intelligent! It is supposed to be not depending
on the media holding contents, for example contents could be in several CDs,
installed on local fixed disk, or located somewhere on the web all using
convenient way for each.
Books can be downloaded from http://thwab.net/books.html.