Import sil-scheherazade-ttf-1.001 as wip/sil-scheherazade-ttf.
Scheherazade, named after the heroine of the classic Arabian Nights tale,
is designed in a similar style to traditional typefaces such as Monotype Naskh,
extended to cover the full Unicode Arabic repertoire.

The goal for this product was to provide a single Unicode-based font family that
would contain a comprehensive inventory of glyphs needed for almost any
Arabic-based writing system. This font makes use of state-of-the-art font
technologies to support complex typographic issues.

This font provides a simplified rendering of Arabic script, using basic
connecting glyphs but not including a wide variety of additional ligatures
or contextual alternates (only the required lam-alef ligatures).
This simplified style is often preferred for clarity, especially in
non-Arabic languages, but may be considered unattractive in more traditional
and literate communities.

One font from this typeface family is included in this release:

    - Scheherazade Regular

This release supports virtually all of the Unicode 5.0 Arabic character
repertoire (excluding the Arabic Presentation Forms blocks, which are not
recommended for normal use). Font smarts are implemented using OpenType
technology.