Import fltk-1.3.x-r7637 as wip/fltk13.
FLTK (pronounced "fulltick") is a LGPL'd C++ graphical user interface
toolkit for X (UNIX), OpenGL, and Win32.  FLTK was designed to be small and
modular enough to be statically linked, but also works fine as a shared
library.  FLTK also includes FLUID, an interactive user interface builder
program.

FLTK 1.3 is the current development branch of the project.  It is
"bleeding edge" and not stable.