As was discussed in pkgsrc-wip-discuss, the cache facility is added. Cache may get fat by repeating (daily or frequent) build. Currently manuall removing of old or all generation is recommended.   GIT_TAG facility is not implemented and to be improved. Thank you.
(1) Converting cvs-package to git-package.
It seems cvs repository is not being updated. So converted to
use git-package.

wip/mk/git-package.mk is recently updated to facilitate caching.
Please take a look for cvs log output for git-package.mk.
Thank you,

(2) update PLIST as of today.


(3) add patches/patch-ad for configure.in, which is not necessary
for regular make, but nice to have further update.