Import qdevelop-0.28 as wip/qdevelop.
QDevelop is a development environment entirely dedicated to Qt4, which requires
Qt4, gcc under Linux or MinGW under Windows, possibly gdb for program debugging
and ctags for code completion.
QDevelop runs on Linux, unix, macosx and windows systems and is released under
the GPL license.

QDevelop is not a Kdevelop like or reduced. It's an independent IDE dedicated to
Qt and is totally independent of KDevelop. Less complete, but faster, light and
especially multi-platforms. QDevelop and KDevelop have different code sources.