Import pyXX-pyexiv2-0.2.2 as wip/py-pyexiv2.
pyexiv2 is a python binding to exiv2, the C++ library for manipulation of EXIF,
IPTC and XMP image metadata. It is a python module that allows your python
scripts to read and write metadata (EXIF, IPTC, XMP, thumbnails) embedded in
image files (JPEG, TIFF, ...).

It is designed as a high-level interface to the functionalities offered by
libexiv2. Using python's built-in data types and standard modules, it provides
easy manipulation of image metadata.

pyexiv2 is distributed under the GPL version 2 license.