Import amanda-client-3.1.2 as wip/amanda3-client.
This is a release of Amanda, the Advanced Maryland Automatic
Network Disk Archiver.  Amanda is a backup system designed to
archive many computers on a network to a single large-capacity
tape drive.

This is Amanda 3.1.2, with perl many advanced features over amanda 2.5.x
in sysutils/amanda-*

The client installation components.