Import amanda-plot-3.1.2nb1 as wip/amanda3-plot.
This is a release of Amanda, the Advanced Maryland Automatic
Network Disk Archiver.  Amanda is a backup system designed to
archive many computers on a network to a single large-capacity
tape drive.

This is Amanda 3.1.2, with perl many advanced features over amanda 2.5.x
in sysutils/amanda-*

This installs the plotting component..