Import py26-xlwt-0.7.2 as wip/py-xlwt.
xlwt is a library for generating spreadsheet files that
are compatible with Excel 97/2000/XP/2003, OpenOffice.org Calc,
and Gnumeric. xlwt has full support for Unicode. Excel
spreadsheets can be generated on any platform without needing
Excel or a COM server. The only requirement is Python 2.3 to 2.6.
xlwt is a fork of pyExcelerator.