Import gupnp-dlna-0.4.0 as wip/gupnp-dlna.
GUPnP is an object-oriented open source framework for creating UPnP devices and
control points, written in C using GObject and libsoup. The GUPnP API is
intended to be easy to use, efficient and flexible.

GUPnP DLNA is a small utility library that aims to ease the DLNA-related tasks
such as media profile guessing, transcoding to a given profile, etc.

GUPnP DLNA is free software released under the GNU LGPL.