Import hs-pathtype-0.5.2 as wip/hs-pathtype.
This package provides type-safe access to filepath manipulations.

System.Path is designed to be used instead of System.FilePath. (It is
intended to provide versions of functions from that module which have
equivalent functionality but are more typesafe). System.Path.Directory
is a companion module providing a type-safe alternative to
System.Directory.

The heart of this module is the Path ar fd abstract type which
represents file and directory paths. The idea is that there are two
phantom type parameters - the first should be Abs or Rel, and the
second File or Dir.