Import xerces-c-3.1.1 as wip/xerces-c.
Xerces-C++ is a validating XML parser written in a portable subset of
C++.  Xerces-C++ makes it easy to give your application the ability to
read and write XML data.  A shared library is provided for parsing,
generating, manipulating, and validating XML documents.

Xerces-C++ is faithful to the XML 1.0 recommendation and many
associated standards.