Import hs-tar-0.3.1.0 as wip/hs-text.
This library is for working with ".tar" archive files. It can read and
write a range of common variations of archive format including V7,
USTAR, POSIX and GNU formats. It provides support for packing and
unpacking portable archives. This makes it suitable for distribution
but not backup because details like file ownership and exact
permissions are not preserved.