Import ec2-ami-tools-1.3.57676 as wip/ec2-ami-tools.
The Amazon EC2 AMI Tools are command-line utilities to help bundle
an Amazon Machine Image (AMI), create an AMI from an existing machine
or installed volume, and upload a bundled AMI to Amazon S3.

TODO: write the modules to create the AMIs. Fairly trivial, the solaris and
linux ones are already provided.