Import py26-tappy-0.8.3 as wip/py-tappy.
TAPPY is a tidal analysis package. It breaks down an hourly record of water
levels into the component sine waves. It is written in Python and uses the least
squares optimization and other functions in SciPy?. The focus is to make the
most accurate analysis possible. TAPPY only determines the constituents
that are calculatable according to the length of the time series