Import ghostscript-9.00 as wip/ghostscript. This is (almost minimum) upgrade from 8.71nb7 and two changes included: (1) options.mk has option: disable-compile-inits (PR/44400) (2) patche-be is added to suppress noisy warning for __func__ duplicate def.
Ghostscript is the well-known PostScript interpreter which is available for
all common and most esoteric platforms and supports many different printers
and some displays.

This is the latest version of Ghostscript, formerly being distributed
under Aladdin Free Public License and released under GPL.  The main
ghostscript version now includes support for cups, enabled by the
"cups" option (which is off by default), and thus it is no longer
necessary to use ghostscript-esp.