Import wqy-zenhei-0.8.38.1 as wip/wqy-zenhei-ttf.
The WenQuanYi Zen Hei is a Hei-Ti style (sans-serif type)
Chinese outline font. It was designed for general purpose text
formatting and on-screen display of Chinese characters among
many other languages. The embolden strokes of the font glyphs
produces enhanced screen contrast, making it easier to read
recognize. The embedded bitmap glyphs further enhance on-screen
performance, which can be enabled with the provided
configuration files.