Initial import of dmd-2.0.51.
TODO:
The compiler itself compiles, but the standard library isn't ported.

Installation hasn't been done yet, so the PLIST is empty.

DESCR:
D is a systems programming language. Its focus is on combining the
power and high performance of C and C++ with the programmer
productivity of modern languages like Ruby and Python. Special
attention is given to the needs of quality assurance, documentation,
management, portability and reliability.

The D language is statically typed and compiles directly to machine
code. It's multiparadigm, supporting many programming styles:
imperative, object oriented, and metaprogramming. It's a member of
the C syntax family, and its appearance is very similar to that of
C++.

It is not governed by a corporate agenda or any overarching theory
of programming. The needs and contributions of the D programming
community form the direction it goes.