Import py26-dana-0.3.1 as wip/py-dana.
DANA (Distributed (asynchronous) Numerical and Adaptive computing) is a python
computing framework based on numpy and scipy libraries whose primary goals
relate to computational neuroscience and artificial neural networks. However,
this framework can be used in several different areas like physic simulations,
cellular automata or image processing.

The computational paradigm supporting the DANA framework is grounded on the
notion of a unit that is a set of arbitrary values that can vary along time
under the influence of other units and learning. Each unit can be linked to any
other unit (including itself) using a weighted link and a group is a structured
set of such homogeneous units.