Import py26-algopy-0.2.2 as wip/py-algopy.
LGOPY is a tool for Algorithmic Differentiation (AD) and
Taylor polynomial approximations. ALGOPY makes it possible to perform
computations on scalar and polynomial matrices. It is designed to be as
compatible to numpy as possible. I.e. views, broadcasting and most functions
of numpy can be performed on polynomial matrices. Exampels are dot,trace,qr,
solve, inv,eigh. The reverse mode of AD is also supported by a simple
code evaluation tracer.