Import py26-mvpa-0.4.5 as wip/py-pymvpa.
PyMVPA is a Python module intended to ease pattern classification
analyses of large datasets. It provides high-level abstraction of
typical processing steps and a number of implementations of some
popular algorithms. While it is not limited to neuroimaging data
it is eminently suited for such datasets.
PyMVPA is truly free software (in every respect) and
additionally requires nothing but free-software to run