Import maloc-1.4 as wip/maloc.
MALOC (Minimal Abstraction Layer for Object-oriented C) is a small, portable,
abstract C environment library for object-oriented C programming. MALOC is used
as the foundation layer for a number of scientific applications, including MC,
SG, and APBS. MALOC can be used as a small stand-alone abstraction
environment for writing portable C programs which need access to resources which
are typically architecture dependent, such as INET sockets, timing routines, and
so on. MALOC provides abstract datatypes, memory management routines, timing
routines, machine epsilon, access to UNIX and INET sockets, MPI, and so on.
All things that can vary from one architecture to another are abstracted out of
an application code and placed in MALOC.