Import liblognorm-0.3.0 as wip/liblognorm.
Briefly described, liblognorm is a tool to normalize log data.

People who need to take a look at logs often have a common problem.
Logs from different machines (from different vendors) usually have
different formats for their logs. Even if it is the same type of log
(e.g. from firewalls), the log entries are so different, that it is
pretty hard to read these. This is where liblognorm comes into the game.
With this tool you can normalize all your logs. All you need is liblognorm
and its dependencies and a sample database that fits the logs you want
to normalize.